categories/serializers.py
from rest_framework import serializers


class CategrySerializer(serializers.Serializer):
    id = serializers.IntegerField(read_only=True)
    name = serializers.CharField(
        required=True,
        max_length=50,
    )
    kind = serializers.CharField(
        max_length=15,
    )
    updated_at = serializers.DateTimeField(read_only=True)
