pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["wesley"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
